+++ src/meltano/cli/discovery.py
@@ -25,13 +25,17 @@ from meltano.core.tracking import GoogleAnalyticsTracker
             PluginType.MODELS,
             PluginType.TRANSFORMS,
             PluginType.ORCHESTRATORS,
-            PluginType.ALL,
+            "all",
         ]
     ),
+    default="all",
 )
 @project()
 def discover(project, plugin_type):
     discover_service = PluginDiscoveryService(project)
+    if plugin_type == "all":
+        plugin_type = None
+
     try:
         discovery_dict = discover_service.discover(plugin_type)