+++ src/meltano/core/config_service.py
@@ -1,11 +1,11 @@
 import os
 import yaml
 import logging
-from typing import Dict, List, Optional
+from typing import Dict, List, Optional, Iterable
 
-from meltano.core.utils import nest
+from meltano.core.utils import nest, NotFound
 from .project import Project
-from .plugin import Plugin, PluginInstall, PluginType, PluginRef
+from .plugin import Plugin, PluginInstall, PluginType, PluginRef, Profile
 from .plugin.factory import plugin_factory
 from .plugin.error import PluginMissingError
 
@@ -18,18 +18,18 @@ class ConfigService:
         os.makedirs(self.project.meltano_dir(), exist_ok=True)
 
     def add_to_file(self, plugin: PluginInstall):
-        installed_def = plugin.canonical()
-
         with self.project.meltano_update() as meltano_yml:
             if not plugin in self.plugins():
-                plugins = nest(meltano_yml, f"plugins.{plugin.type}", value=[])
-                plugins.append(installed_def)
+                if not plugin.type in meltano_yml.plugins:
+                    meltano_yml.plugins[plugin.type] = []
+
+                meltano_yml.plugins[plugin.type].append(plugin)
             else:
                 logging.warning(
                     f"{plugin.name} is already present, use `meltano install` to install it."
                 )
 
-        return plugin_factory(plugin.type, installed_def)
+        return plugin_factory(plugin.type, plugin.canonical())
 
     def has_plugin(self, plugin_name: str):
         try:
@@ -39,7 +39,7 @@ class ConfigService:
             return False
 
     def find_plugin(self, plugin_name: str, plugin_type: Optional[PluginType] = None):
-        name, profile = PluginRef.parse_name(plugin_name)
+        name, profile_name = PluginRef.parse_name(plugin_name)
         try:
             plugin = next(
                 plugin
@@ -50,10 +50,29 @@ class ConfigService:
                 )
             )
 
-            plugin.profile = profile
+            profile = self.find_plugin_profile(plugin, profile_name)
+            plugin.use_profile(profile)
+
             return plugin
         except StopIteration as stop:
             raise PluginMissingError(name) from stop
+
+    def get_plugin(self, plugin_ref: PluginRef) -> PluginInstall:
+        try:
+            plugin = next(plugin for plugin in self.plugins() if plugin == plugin_ref)
+
+            profile = self.find_plugin_profile(plugin, plugin_ref.current_profile_name)
+            plugin.use_profile(profile)
+
+            return plugin
+        except StopIteration as stop:
+            raise PluginMissingError(plugin_ref.name) from stop
+
+    def find_plugin_profile(self, plugin, profile_name) -> Optional[Profile]:
+        try:
+            return plugin.get_profile(profile_name)
+        except NotFound:
+            return Profile.DEFAULT
 
     def get_extractors(self):
         return filter(lambda p: p.type == PluginType.EXTRACTORS, self.plugins())
@@ -70,20 +89,19 @@ class ConfigService:
     def get_models(self):
         return filter(lambda p: p.type == PluginType.MODELS, self.plugins())
 
-    def get_database(self, database_name):
-        return yaml.load(
-            open(self.project.meltano_dir(f".database_{database_name}.yml"))
-        )
+    def update_plugin(self, plugin: PluginInstall):
+        with self.project.meltano_update() as meltano:
+            # find the proper plugin to update
+            idx, outdated = next(
+                (i, it)
+                for i, it in enumerate(meltano.plugins[plugin.type])
+                if it == plugin
+            )
 
-    def plugins(self) -> List[PluginInstall]:
-        """Parse the meltano.yml file and return it as `PluginInstall` instances."""
+            meltano.plugins[plugin.type][idx] = plugin
 
-        # this will parse the meltano.yml file and create an instance of the
-        # corresponding `plugin_class` for all the plugins.
-        return (
-            plugin_factory(plugin_type, plugin_def)
-            for plugin_type, plugin_defs in self.project.meltano.get(
-                "plugins", {}
-            ).items()
-            for plugin_def in plugin_defs
-        )
+            return outdated
+
+    def plugins(self) -> Iterable[PluginInstall]:
+        for plugin_type, plugins in self.project.meltano.plugins:
+            yield from plugins