+++ src/webapp/src/components/pipelines/ConnectorSettings.vue
@@ -91,21 +91,33 @@ export default {
         this.snowflakeHelper(newVal)
       },
       deep: true
+    },
+    'configSettings.profileInFocusIndex': {
+      handler(newVal, oldVal) {
+        this.refocusInput(newVal, oldVal)
+      }
     }
   },
   mounted() {
-    this.$nextTick(this.focusInputIntelligently)
+    this.focusInputIntelligently()
   },
   methods: {
     findLabel(setting) {
       return setting.options.find(item => item.value === setting.value).label
     },
     focusInputIntelligently() {
-      const inputs = Array.from(this.$el.getElementsByTagName('input'))
-      if (inputs.length) {
-        const firstEmptyInput = inputs.find(el => !el.value)
-        const targetInput = firstEmptyInput || inputs[0]
-        targetInput.focus()
+      this.$nextTick(() => {
+        const inputs = Array.from(this.$el.getElementsByTagName('input'))
+        if (inputs.length) {
+          const firstEmptyInput = inputs.find(el => !el.value)
+          const targetInput = firstEmptyInput || inputs[0]
+          targetInput.focus()
+        }
+      })
+    },
+    refocusInput(newVal, oldVal) {
+      if (newVal !== oldVal) {
+        this.focusInputIntelligently()
       }
     },
     snowflakeHelper(newVal) {
@@ -117,7 +129,8 @@ export default {
        * when/if we add UX helpers like this for more connectors
        * TODO: Need to add a loader indicator to show something is "processing"
        */
-      const accountInput = newVal.config.account
+      const accountInput =
+        newVal.profiles[newVal.profileInFocusIndex].config.account
       if (accountInput) {
         const parsedAccountId = utils.snowflakeAccountParser(accountInput)
 
@@ -125,10 +138,14 @@ export default {
           const vm = this
 
           setTimeout(() => {
-            vm.configSettings.config.account = parsedAccountId
+            vm.configSettings.profiles[
+              vm.configSettings.profileInFocusIndex
+            ].account = parsedAccountId
           }, 1000)
         } else {
-          this.configSettings.config.account = newVal.config.account
+          this.configSettings.profiles[
+            this.configSettings.profileInFocusIndex
+          ].account = newVal.profiles[newVal.profileInFocusIndex].config.account
         }
       }
     }
@@ -163,7 +180,10 @@ export default {
               <input
                 v-if="getIsOfKindBoolean(setting.kind)"
                 :id="getFormFieldForId(setting)"
-                v-model="configSettings.config[setting.name]"
+                v-model="
+                  configSettings.profiles[configSettings.profileInFocusIndex]
+                    .config[setting.name]
+                "
                 class="checkbox"
                 :class="successClass(setting)"
                 type="checkbox"
@@ -172,7 +192,10 @@ export default {
               <!-- Date -->
               <InputDateIso8601
                 v-else-if="getIsOfKindDate(setting.kind)"
-                v-model="configSettings.config[setting.name]"
+                v-model="
+                  configSettings.profiles[configSettings.profileInFocusIndex]
+                    .config[setting.name]
+                "
                 :name="setting.name"
                 :for-id="getFormFieldForId(setting)"
                 :input-classes="`is-small ${successClass(setting)}`"
@@ -185,7 +208,10 @@ export default {
               >
                 <select
                   :id="`${setting.name}-select-menu`"
-                  v-model="configSettings.config[setting.name]"
+                  v-model="
+                    configSettings.profiles[configSettings.profileInFocusIndex]
+                      .config[setting.name]
+                  "
                   :name="`${setting.name}-options`"
                   :class="successClass(setting)"
                 >
@@ -204,7 +230,10 @@ export default {
               <input
                 v-else-if="getIsOfKindTextBased(setting.kind)"
                 :id="getFormFieldForId(setting)"
-                v-model="configSettings.config[setting.name]"
+                v-model="
+                  configSettings.profiles[configSettings.profileInFocusIndex]
+                    .config[setting.name]
+                "
                 :class="['input', fieldClass, successClass(setting)]"
                 :type="getTextBasedInputType(setting)"
                 :placeholder="setting.value || setting.name"