+++ src/webapp/src/components/pipelines/CreateScheduleModal.vue
@@ -34,12 +34,27 @@ export default {
   },
   computed: {
     ...mapState('configuration', ['recentELTSelections', 'transformOptions']),
-    ...mapGetters('plugins', ['getHasInstalledPluginsOfType']),
+    ...mapGetters('plugins', [
+      'getHasInstalledPluginsOfType',
+      'getPluginProfiles'
+    ]),
     ...mapState('plugins', ['installedPlugins']),
     isSaveable() {
       const hasOwns = []
       _.forOwn(this.pipeline, val => hasOwns.push(val))
       return hasOwns.find(val => val === '' || val === null) === undefined
+    },
+    availableExtractors() {
+      return _(this.installedPlugins.extractors)
+        .map(this.getPluginProfiles)
+        .flatten()
+        .value()
+    },
+    availableLoaders() {
+      return _(this.installedPlugins.loaders)
+        .map(this.getPluginProfiles)
+        .flatten()
+        .value()
     }
   },
   created() {
@@ -146,9 +161,9 @@ export default {
                       :disabled="!getHasInstalledPluginsOfType('extractors')"
                     >
                       <option
-                        v-for="extractor in installedPlugins.extractors"
-                        :key="extractor.name"
-                        >{{ extractor.name }}</option
+                        v-for="extractor in availableExtractors"
+                        :key="extractor"
+                        >{{ extractor }}</option
                       >
                     </select>
                   </span>
@@ -166,9 +181,9 @@ export default {
                       :disabled="!getHasInstalledPluginsOfType('loaders')"
                     >
                       <option
-                        v-for="loader in installedPlugins.loaders"
-                        :key="loader.name"
-                        >{{ loader.name }}</option
+                        v-for="loader in availableLoaders"
+                        :key="loader"
+                        >{{ loader }}</option
                       >
                     </select>
                   </span>