+++ src/webapp/src/components/pipelines/LoaderSettingsModal.vue
@@ -6,12 +6,14 @@ import lodash from 'lodash'
 
 import ConnectorLogo from '@/components/generic/ConnectorLogo'
 import ConnectorSettings from '@/components/pipelines/ConnectorSettings'
+import ConnectorSettingsDropdown from '@/components/pipelines/ConnectorSettingsDropdown'
 
 export default {
   name: 'LoaderSettingsModal',
   components: {
     ConnectorLogo,
-    ConnectorSettings
+    ConnectorSettings,
+    ConnectorSettingsDropdown
   },
   data() {
     return {
@@ -22,7 +24,8 @@ export default {
     ...mapGetters('plugins', [
       'getInstalledPlugin',
       'getIsPluginInstalled',
-      'getIsInstallingPlugin'
+      'getIsInstallingPlugin',
+      'getPluginProfiles'
     ]),
     ...mapGetters('configuration', ['getHasValidConfigSettings']),
     ...mapState('configuration', ['loaderInFocusConfiguration']),
@@ -36,15 +39,24 @@ export default {
     isLoadingConfigSettings() {
       return !Object.prototype.hasOwnProperty.call(
         this.localConfiguration,
-        'config'
+        'profiles'
       )
     },
     isSaveable() {
+      if (this.isInstalling || this.isLoadingConfigSettings) {
+        return
+      }
+      const configSettings = {
+        config: this.localConfiguration.profiles[
+          this.localConfiguration.profileInFocusIndex
+        ].config,
+        settings: this.localConfiguration.settings
+      }
       const isValid = this.getHasValidConfigSettings(
-        this.localConfiguration,
+        configSettings,
         this.loader.settingsGroupValidation
       )
-      return !this.isInstalling && this.isInstalled && isValid
+      return this.isInstalled && isValid
     },
     loader() {
       return this.getInstalledPlugin('loaders', this.loaderName)
@@ -82,7 +94,7 @@ export default {
     },
     createEditableConfiguration() {
       this.localConfiguration = Object.assign(
-        {},
+        { profileInFocusIndex: 0 },
         lodash.cloneDeep(this.loaderInFocusConfiguration)
       )
     },
@@ -97,7 +109,7 @@ export default {
         .dispatch('configuration/savePluginConfiguration', {
           name: this.loader.name,
           type: 'loaders',
-          config: this.localConfiguration.config
+          profiles: this.localConfiguration.profiles
         })
         .then(() => {
           this.$store.dispatch('configuration/updateRecentELTSelections', {
@@ -143,6 +155,11 @@ export default {
               <a :href="loader.signupUrl" target="_blank">sign up here</a>.
             </p>
           </div>
+          <ConnectorSettingsDropdown
+            :connector="loader"
+            plugin-type="loaders"
+            :config-settings="localConfiguration"
+          ></ConnectorSettingsDropdown>
           <ConnectorSettings
             field-class="is-small"
             :config-settings="localConfiguration"